Add the ability to exit a game.
If any player exits the game the game server will be cleared. Otherwise
the player will still be in the game when then login again. If all
players simply close the browser and don't exit the game will remain and
this needs to be addressed with some cleanup of abandoned games.